+++ tests/mysql/table.rs
@@ -439,9 +439,9 @@ fn create_with_named_check_constraint() {
                 ColumnDef::new(Glyph::Id)
                     .integer()
                     .not_null()
-                    .check_with_name("positive_id", Expr::col(Glyph::Id).gt(10))
+                    .check(("positive_id", Expr::col(Glyph::Id).gt(10)))
             )
-            .check_with_name("id_range", Expr::col(Glyph::Id).lt(20))
+            .check(("id_range", Expr::col(Glyph::Id).lt(20)))
             .check(Expr::col(Glyph::Id).ne(15))
             .to_string(MysqlQueryBuilder),
         r"CREATE TABLE `glyph` ( `id` int NOT NULL CONSTRAINT `positive_id` CHECK (`id` > 10), CONSTRAINT `id_range` CHECK (`id` < 20), CHECK (`id` <> 15) )",
@@ -475,7 +475,7 @@ fn alter_with_named_check_constraint() {
                     .integer()
                     .not_null()
                     .default(101)
-                    .check_with_name("positive_aspect", Expr::col(Glyph::Aspect).gt(100))
+                    .check(("positive_aspect", Expr::col(Glyph::Aspect).gt(100)))
             )
             .to_string(MysqlQueryBuilder),
         r#"ALTER TABLE `glyph` ADD COLUMN `aspect` int NOT NULL DEFAULT 101 CONSTRAINT `positive_aspect` CHECK (`aspect` > 100)"#,