+++ CHANGELOG.md
@@ -28,6 +28,8 @@ pub struct SeaRc;                                 // new
 
 ### Breaking Changes
 
+* Replace `ColumnSpec::Check(Expr)` with `ColumnSpec::Check(Check)` to support named check constraints
+
 * Removed inherent `SimpleExpr` methods that duplicate `ExprTrait`. If you encounter the following error, please add `use sea_query::ExprTrait` in scope https://github.com/SeaQL/sea-query/pull/890
 ```rust
 error[E0599]: no method named `like` found for enum `sea_query::Expr` in the current scope
@@ -124,6 +126,7 @@ impl Iden for Glyph {
 
   If you had custom implementations in your own code, some may no longer compile
   and may need to be deleted.
+
 
 ### Upgrades